refactor(react-kit): move tx error parsing to a shared util
